requirements.txt
flask==3.0.3
gTTS==2.5.1
newspaper3k==0.2.8
pydub==0.25.1
lxml[html_clean]==5.3.0
gunicorn==22.0.0
